+++ controllers/workController.js
@@ -11,4 +11,15 @@ const worksList = async (req, res) => {
   }
 };
 
-module.exports = { worksList };
+// 지정된 피드 상세
+const feed = async (req, res) => {
+  try {
+    const result = await workService.feed();
+    res.status(200).json(result);
+  } catch (err) {
+    console.log(err);
+    res.status(err.statusCode).json({ message: err.message });
+  }
+};
+
+module.exports = { worksList, feed };
