+++ services/feedService.js
@@ -1,14 +1,13 @@
 const feedDao = require('../models/feedDao');
 
-// 지정된 피드 상세
-const feed = async () => {
+// 최신 feed list
+const feedsList = async () => {
   try {
-    const result = await feedDao.feed();
+    const result = await feedDao.feedsList();
     return result;
   } catch (err) {
     console.log(err);
     res.status(err.statusCode).json({ message: err.message });
   }
 };
-
-module.exports = { feed };
+module.exports = { feedsList };
